Fix various warnings raised by GCC 5

+++ core/node_path.cpp
@@ -276,7 +276,7 @@ NodePath NodePath::get_as_property_path() const {
 
 		String initial_subname = data->path[0];
 
-		for (size_t i = 1; i < data->path.size(); i++) {
+		for (int i = 1; i < data->path.size(); i++) {
 			initial_subname += "/" + data->path[i];
 		}
 		new_path.insert(0, initial_subname);
